SendFunds now sending all options to Dwolla
SendTransactionOptions was updated some time ago to support the latest api from dwolla. The DwollaTransactionService was not updated to reflect these newly added properties. This commit fixes any issues related to sending funds with unintended consequences. Specifically, the funding source - which was causing all sorts of chaos for my project.